Document-AI-Hub/requirements.txt
accelerate==1.13.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioredis==2.0.1
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.11.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
async-timeout==5.0.1
asyncer==0.0.17
asyncpg==0.31.0
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bidict==0.23.1
billiard==4.2.4
build==1.4.2
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
chainlit==2.10.0
charset-normalizer==3.4.6
chevron==0.14.0
chromadb==1.5.5
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
colorlog==6.10.1
cryptography==46.0.6
cuid==0.4
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.3.1
diff-match-patch==20241021
dill==0.4.1
distro==1.9.0
docling==2.82.0
docling-core==2.70.2
docling-ibm-models==3.12.0
docling-parse==5.6.1
docstring_parser==0.17.0
docx==0.2.4
durationpy==0.10
elastic-transport==9.2.1
elasticsearch==9.3.0
et_xmlfile==2.0.0
faiss-cpu==1.13.2
Faker==37.12.0
fastapi==0.135.2
fastuuid==0.14.0
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
guardrails-ai==0.9.3
guardrails-api-client==0.4.0
guardrails_hub_types==0.0.4
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.3.0
isoduration==20.11.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
kubernetes==35.0.0
langchain==0.0.300
langchain-core==1.2.22
langsmith==0.7.22
lark==1.3.1
latex2mathml==3.79.0
Lazify==0.4.0
lazy_imports==1.2.0
litellm==1.82.6
literalai==0.1.201
llvmlite==0.46.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.26.0
mdurl==0.1.2
minio==7.2.20
mmh3==5.2.1
more-itertools==10.8.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numba==0.64.0
numexpr==2.14.1
numpy==2.4.3
oauthlib==3.3.1
ocrmac==1.0.1
omegaconf==2.3.0
onnxruntime==1.24.4
openai==2.29.0
openai-whisper==20250625
opencv-python==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-agno==0.53.3
opentelemetry-instrumentation-alephalpha==0.53.3
opentelemetry-instrumentation-anthropic==0.53.3
opentelemetry-instrumentation-bedrock==0.53.3
opentelemetry-instrumentation-chromadb==0.53.3
opentelemetry-instrumentation-cohere==0.53.3
opentelemetry-instrumentation-crewai==0.53.3
opentelemetry-instrumentation-google-generativeai==0.53.3
opentelemetry-instrumentation-groq==0.53.3
opentelemetry-instrumentation-haystack==0.53.3
opentelemetry-instrumentation-lancedb==0.53.3
opentelemetry-instrumentation-langchain==0.53.3
opentelemetry-instrumentation-llamaindex==0.53.3
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-instrumentation-marqo==0.53.3
opentelemetry-instrumentation-mcp==0.53.3
opentelemetry-instrumentation-milvus==0.53.3
opentelemetry-instrumentation-mistralai==0.53.3
opentelemetry-instrumentation-ollama==0.53.3
opentelemetry-instrumentation-openai==0.53.3
opentelemetry-instrumentation-openai-agents==0.53.3
opentelemetry-instrumentation-pinecone==0.53.3
opentelemetry-instrumentation-qdrant==0.53.3
opentelemetry-instrumentation-redis==0.61b0
opentelemetry-instrumentation-replicate==0.53.3
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-instrumentation-sagemaker==0.53.3
opentelemetry-instrumentation-sqlalchemy==0.61b0
opentelemetry-instrumentation-threading==0.61b0
opentelemetry-instrumentation-together==0.53.3
opentelemetry-instrumentation-transformers==0.53.3
opentelemetry-instrumentation-urllib3==0.61b0
opentelemetry-instrumentation-vertexai==0.53.3
opentelemetry-instrumentation-voyageai==0.53.3
opentelemetry-instrumentation-watsonx==0.53.3
opentelemetry-instrumentation-weaviate==0.53.3
opentelemetry-instrumentation-writer==0.53.3
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-semantic-conventions-ai==0.4.16
opentelemetry-util-http==0.61b0
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==3.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pluggy==1.6.0
polyfactory==3.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
psycopg2-binary==2.9.11
pybase64==1.4.3
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydash==8.0.6
Pygments==2.19.2
PyJWT==2.12.1
pylatexenc==2.10
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreML==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Vision==12.1
pypdf==6.9.2
pypdfium2==5.6.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-engineio==4.13.1
python-multipart==0.0.22
python-pptx==1.0.2
python-socketio==5.16.1
PyYAML==6.0.3
rapidocr==3.7.0
redis==7.4.0
referencing==0.37.0
regex==2026.2.28
requests==2.33.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
rstr==3.2.2
rtree==1.4.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semchunk==3.2.5
semver==3.0.4
sentence-transformers==5.3.0
sentencepiece==0.2.1
setuptools==81.0.0
shapely==2.1.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.20
sse-starlette==3.3.3
starlette==1.0.0
sympy==1.14.0
syncer==2.0.3
tabulate==0.10.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.1
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
traceloop-sdk==0.53.3
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
webcolors==25.10.0
websocket-client==1.9.0
websockets==16.0
wheel==0.46.3
whisper==1.1.10
wrapt==1.17.3
wsproto==1.3.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
